06/03/2025 final

+++ src/main/java/com/ghoul/AmanaFund/service/ISinistreService.java
@@ -21,5 +21,5 @@ public interface ISinistreService {
 
     double predireFondsDeReserve();
 
-    int evaluerRisque(Long userId);
+    int evaluerRisque(int userId);
 }
